[DREAM-813] Move tooltip caret placement to shared
The timeline tooltip's caret geometry is generic: it only needs the
popover and anchor rectangles. Moves it under a shared anchored-popover
directory so the calendar and budget chart tooltips can use it too.

https://community.openproject.org/wp/DREAM-813
