+++ frontend/src/app/shared/components/anchored-popover/caret-placement.ts
